Update scatter mechanism
The new scatter mechanism seems like a more sensible choice that simply adding things together in the data reservoir.
If the selection weight after the dimension increasing random projection is less than zero no change is made to that element in the reservoir.  If it is between zero and one the new value is a weighted blend between the old value and the new.  Zero leaving the old value in place.  One completely replacing the old value.  All the weight value are always between minus one and one.
Also added (self) gated associative memory via truncation.